move core implementation to generators/base-core
